atualizacao do agendamento e de outros componentes do aplicativo

+++ vitrinebeauty/lib/model/model_card_historico.dart
@@ -1,6 +1,6 @@
 class ModelCardHistorico {
   String? id;
-  int? avaliacao;
+  double? avaliacao;
   String? data;
   String? hora;
   String? nomeCliente;
